+++ src/sinks.nim
@@ -351,6 +351,13 @@ proc getSinkConfigByName*(name: string): Option[SinkConfig] =
         else:
           # Nimutils wants this param as a string.
           opts[k] = $(unpack[int](boxOpt.get()))
+    of "disable_after_errors":
+      let boxOpt = attrGetOpt[Box](section & "." & k)
+      if boxOpt.isSome():
+        if boxOpt.get().kind != MkInt:
+          error(k & " (sink config key) must be a positive integer")
+        else:
+          opts[k] = $(unpack[int](boxOpt.get()))
     of "max":
       try:
         # Todo: move this check to a type check in the spec.